requirements.txt
# hunt.py, scrape, score, filter
python-jobspy>=1.1
pypdf>=3.0
pandas>=2.0
openpyxl>=3.1
pdfplumber>=0.10
rapidfuzz>=3.0
sentence-transformers>=2.0
# deepseek_ats_analyzer
requests>=2.28
# api/app.py
flask>=3.0
flask-cors>=4.0